detect risc-v vector extension
The RISC-V Vector Extension adds a vcsr CSR since version 0.8, which
does not appear in version 0.7.1. I confirmed this on Sipeed Lichee
Pi 4A, which has T-Head C910 cores inside, and the version of RVV is
0.7.1, Checking the existence of the vlenb CSR will be a way to
probe RVV but not v0p7.

For some concerns that some older kernels do not enable vector
support for userspace, this solution will also work to treat it as
the vector extension does not exist. This can be confirmed on T-Head
Vendor Kernel working on TH1520. As the RISC-V Vector Extension
Specification said: "Attempts to execute any vector instruction,
or to access the vector CSRs, raise an illegal-instruction exception
when mstatus.VS is set to Off."

Signed-off-by: Yangyu Chen <cyy@cyyself.name>